routes: Add API to grab pipeline status
